Intigrate Migration into UnrealEngine Blueprints (#152)
Add migratable snapshots to the UE engine and add the new AsyncAction
CreateMigratableSnapshot in UE and add it to the demo project.

Also fixes in the migration logic:
+ support decreasing knot count
+ The list_table migration was broken
+ List similarity was calculated inverse
+ Migrate lists in linked-in variables
+ Loaded tags are now stored in string_table

Bump Doxygen 1.17
Bump GitHub actions to NodeJS 24

Python interface:
+ Add `lookaheadSafe` to API

+++ inkcpp_compiler/emitter.cpp
@@ -8,7 +8,7 @@
 
 namespace ink::compiler::internal
 {
-void emitter::start(int ink_version, compilation_results* results)
+void emitter::start(uint16_t ink_version, compilation_results* results)
 {
 	// store
 	_ink_version = ink_version;
